feat: separe types + add webpack for svg
