+++ faststore.config.js
@@ -99,9 +99,20 @@ module.exports = {
 
   webpack(config) {
     config.module.rules.push({
-      test: /\.svg$/,
-      use: ['@svgr/webpack'],
+      ...fileLoaderRule,
+      test: /\.svg$/i,
+      resourceQuery: /url/,
+    },
+
+    {
+      test: /\.svg$/i,
+      issuer: fileLoaderRule.issuer,
+      resourceQuery: { not: [...fileLoaderRule.resourceQuery.not, /url/] },
+      use: ["@svgr/webpack"],
     });
+
+    fileLoaderRule.exclude = /\.svg$/i;
+    
     return config;
   },
 }
