+++ src/components/BannerGrid/BannerGridTypes.d.ts
@@ -1,0 +1,15 @@
+interface banner {
+  image: string;
+  name: string;
+  url: string;
+  width: string;
+}
+
+interface column {
+  column: banner[];
+}
+
+
+export interface bannerArray {
+  row: column[];
+}
