+++ src/components/CustomBreadcrumb/CustomBreadcrumb.tsx
@@ -5,14 +5,9 @@ import { usePDP }  from "@faststore/core"
 
 import { Breadcrumb } from '@faststore/ui'
 
-import SVGIMG from '!@svgr/webpack!../../assets/accountIcon.svg';
+import { CustomBreadcrumbProps } from './CustomBreadcrumbTypes'
 
-
-
-export interface CustomBreadcrumbProps {
-  showSku: boolean;
-  color: string;
-}
+//import SVGIMG from '!@svgr/webpack!../../assets/accountIcon.svg';
 
 export default function CustomBreadcrumb(props: CustomBreadcrumbProps) {
 
@@ -26,7 +21,6 @@ export default function CustomBreadcrumb(props: CustomBreadcrumbProps) {
     return null
   }
 
-
   if(props?.showSku) {
     breadCrumb[breadCrumb.length-1] = {
       item: '#',
@@ -37,7 +31,6 @@ export default function CustomBreadcrumb(props: CustomBreadcrumbProps) {
   
   return (
     <section className={styles.CustomBreadcrumb} >
-        <SVGIMG fill="red" width={222} />
       <Breadcrumb breadcrumbList={breadCrumb} />
     </section>
   );
