+++ src/components/CustomBreadcrumb/CustomBreadcrumbTypes.d.ts
@@ -1,0 +1,4 @@
+export interface CustomBreadcrumbProps {
+  showSku: boolean;
+  color: string;
+}
