+++ src/components/ImageRow/imageRowTypes.d.ts
@@ -1,0 +1,12 @@
+interface List {
+  image: string;
+  name: string;
+  url: string;
+}
+export interface ImageRowProps {
+  list: List[];
+  title: string;
+  imageHeight: string;
+  itemsPerPage: string;
+  titleEnabled: string;
+}
