+++ src/components/InfoCardSlider/InfoCardSliderTypes.d.ts
@@ -1,0 +1,12 @@
+interface List {
+  imageDesktop: string;
+  name: string;
+  url: string;
+}
+export interface InfoCardSliderProps {
+  list: List[];
+  itemsPerPage: number;
+  infiniteMode: boolean;
+  title: string;
+  titleEnabled: boolean;
+}
