+++ src/components/ManufacturerNumber/ManufacturerNumber.tsx
@@ -3,18 +3,12 @@ import styles from './ManufacturerNumber.module.scss'
 
 import { usePDP }  from "@faststore/core"
 
-
-
-
 export default function ManufacturerNumber() {
 
 
   const context = usePDP()
 
   const properties = context?.data?.product?.properties;
-
-  console.log(context)
-
 
   let ManufacturerNumber = properties?.find((spec: any) => spec.name === "Manufacturer Part Number")?.values?.[0].val
