+++ src/components/ProductSpecifications/ProductSpecificationsTypes.d.ts
@@ -1,0 +1,3 @@
+export interface ProductSpecificationsList {
+  headTitle: string;
+}
