+++ src/components/TabContent/TabContentTypes.d.ts
@@ -1,0 +1,15 @@
+interface banner {
+  image: string;
+}
+
+interface tabs {
+  title: string;
+  banner: banner;
+  content: string;
+  linkText: string;
+  href: string;
+  
+}
+export interface TabContentProps {
+  tabs: tabs[];
+}
